+++ .vscode/tasks.json
@@ -13,6 +13,7 @@
 				"arduino/command.cpp",
                 "arduino/serial.cpp",
                 "arduino/pin_control.cpp",
+                "arduino/button_control.cpp",
 				//dmyコード
 				"Sim/dmy_Serial.cpp",
 				"Sim/dmy_Arduino.cpp",
