+++ arduino/arduino.ino
@@ -2,6 +2,7 @@
 
 #include "serial.h"
 #include "pin_control.h"
+#include "button_control.h"
 
 // ボタンが押されたときの処理(仮)
 void onButtonPress() {
@@ -13,12 +14,12 @@ void setup()
   SerialInit();
 
   // ボタン用ピンの初期化（7番ピンにコールバックを登録）
-  PinControlButtonInit(7, onButtonPress);
+  ButtonInit(7, onButtonPress);
 }
 
 void loop()
 {
   SerialRead();
-  PinControlButtonRead();
+  ButtonRead();
 }
