+++ arduino/button_control.h
@@ -1,0 +1,9 @@
+#ifndef BUTTON_CONTROL_H
+#define BUTTON_CONTROL_H
+
+typedef void (*ButtonCallback)(void);
+
+unsigned short ButtonInit(unsigned short pin, ButtonCallback callback);
+unsigned short ButtonRead(void);
+
+#endif // BUTTON_CONTROL_H
